Move serializer into lib/console.
