+++ lib/console/serializer.ts
